database,web/auth: Collect all database methods under an interface
The initial idea was to have multiple smaller interfaces for users, courses,
etc. Thus, handler functions could take a UserDatabase interface if only
modifying a user. I'm changing this back as it might be premature, and not even
needed. For now I'll just collect the methods under a common Database interface.