web/courses_test.go: Fix broken test
227ad6501b0afb719f7a3e3ccdab8177779d8a5e introduced automatic course enrollment
for teachers creating a new course. The teacher's user id now needs to be
extracted from the user session, and to create the enrollment that user must
also exist in the database.

The user is extracted from the teacher
session, thus a user must be present in the session during the test.