web/courses.go,database/gormdb.go: support detection of duplicate course creation...
This commit adds support for detecting if a course was already created
in the database for a directoryID (organization).
It also allows repo creation to fail and continue and thus create
hooks for existing repos. This is useful if a teacher already
created one or more of the repos, yet have the necessary hooks
established for the course.