refactored database.CreateCourse() to take course creator (user)
This change makes it explicit who is the course creator by
checking that the user is admin and should be allowed to create
the course. Further, the user is then enrolled in the course
as teacher.

This change impacted a lot of tests, but almost no business logic.
It basically only changes web/courses.go by moving the enrollment
logic into the database.CreateCourse() method, since this should
actually be done in one transaction, so that if the user couldn't
be enrolled as teacher, the course won't created either.
Currently, the implementation does not support transactions, but
it should be easy to do with GORM at a later time, if necessary.

All the other changes are mostly related to adapting to the new
semantics of CreateCourse(), but there are also some simplifications
to some of the tests, which should be a starting point for test
complexity reduction...