refactor db API; require withRemoteID in GetUsers() and GetGroup()
This commit adds a parameter to GetUsers() and GetGroup() to allow
specifying that the users returned and the group member users
should be returned with their corresponding remote identities
(if the withRemoteID parameter is true). Otherwise the remote
identities will not be included in the user data sets.

Important: the remote identities should not be needed in the
frontend, and so data returned to the frontend should use these
methods with 'false' for the withRemoteID, so as to return users
without the remote identities.