fixed bug in UpdateGroup() and some refactoring; added test
This commit fixes a bug in UpdateGroup() that prevented teachers
from approving groups and at the same time adding/removing members.
We now update the group in the database first, before doing
various checks for enrollment etc. Only when these check have
passed do we create the group repository on the SCM and add
the repository to the database.

Note that we set UserID to 0 for group repositories in the database
(hence the corresponding changes in both UpdateGroup and PatchGroup)
and we set GroupID instead. Now we also add the HTMLURL to the
repository in the database.

This commit also adds TestNewGroupStudentCreateGroupWithTeacher
to test that a student can create a group and two different
teachers can change the group (one adds one user and the other
removes two users.)