super major refactoring of ci, assignment refresh and parsing
This commit is pretty big, but it was hard to do incrementally.
Mainly, the commit aims to improve code readability, structure
and testability. There are some TODOs remaining and there are
some fmt.Print and t.Logf() calls that should be removed when
tests are made proper.

Some (compatability) notes:
- copied and edited go.tmpl and python361.tmpl to ci/scripts
  (we may want to keep only one of these...)
- there should not be any changes to the database schema
- the assignments returned from web.RefreshCourse() is now
  models.Assignment instead of yamlparser.NewAssignmentRequest;
  however, the frontend does not seem to use the data returned
  from RefreshCourse().
- yamlparser.NewAssignmentRequest should not be used outside
  this package, and so the yamlparser.Parse() function instead
  returns []models.Assignment.
- we can probably discard ciresult.go, since the structs that
  are needed have been moved to either ci package or are fetched
  from github.com/autograde/kit/score package.

There are also many changes in web/hooks.go and elsewhere.