made test runner use random secret per execution
This commit replaces common/secret.go on each test run with
a custom one with a random value for the Secret().
This secret is then available for result extraction after the
test run. Moreover, if we detect attempts at trying to use
the secret in student code, we fail the tests and exit without
running the test itself.