experimenting with go.uber.org/zap logger
This commit is mainly an experiment to better understand if Uber's
zap logger can replace logrus. It is supposed to be faster and
support for both development (debugging) and production logging
is easy to configure.

I'm committing this separately from vendoring the dependencies
to avoid messing up the diff.

To use this logger you need to do:
  go get -u go.uber.org/zap