removed a panic and added some todos in gorm.db
This commit mainly removes an unncessary panic in setEnrollment()
This is a package local method, and is currently only used from
Enroll* methods, which use the predefined enum types. Moreover,
the type of the status argument also prevents any inputs from being
outside the domain of the enum type.

Otherwise, this commit adds a few TODO items.