added logger to web/auth package
Now all functions that do auth takes a logger and logs all errors
and a few info elements.

PS: I have not tested this to see the volum of log messages.
Some of them may be turned into Debug messages or removed
altogether if they generate significant message volum.
(some error paths may represent expected behavior and so may
not need to be logged, but for now we log every error message.)