added logger to NewAutograderService and tweaked error/logging
This commit is the first attempt to understand how to best
deal with errors and logging. The principle is that all logging
of returned errors take place in AutograderService methods.
Furthermore, the AutograderService methods are the ones that
return status codes and error strings to the frontend, so as
to prevent server internal error details from being exposed
in the frontend; devops will have access to server logs for
debugging purposes.

PS: This should be considered experimental for now, while we
assess the merit of this approach.