consolidated several functions in the database interface
This commit removes several functions from the database interface
which had overlapping implementation, but different inputs.
We now instead provide a single GetRepositories() that only takes
a pb.Repositories{} as a query parameter that can be passed to
the database directly.

PS: When using this, we need to be careful to not pass information
from the frontend (external) directly to the database query, but
instead create a new query object to be used for the specific
gRPC function.