migrating grpcservice package into web and resolved import cycle
This is an attempt to migrate the grpcservice package into the
web package to allow simpler access to various fields of the
AutograderService struct, and thereby avoid passing these to
the web.Methods(). The goal is also to clean up the error handling
and return only the status error to the frontend from one place
in the autograder_service.go file.