refactored to use interceptor for message argument validation
This commit refactors the input message argument checks to use
a common 'validator' interface with a single IsValid() method.
This allows a single interceptor implementation ensuring that
all methods call their respective IsValid() method.
If no such method exists for some message type, the interceptor
will log a debug message. This is not a bug per se but we do
want to implement an IsValid() method for all message types.

Note that this required changing the methods
CreateEnrollment() and UpdateEnrollment() to use the
Enrollment message type instead of ActionRequest to avoid
collision between two different IsValid() methods previously
implemented on the ActionRequest.

Also note that the DeleteGroup() method currently will invoke
the Group.IsValid() method, which was not previously done.
This problem is not yet resolved, but there is a TODO item
that explains a solution for this.

The frontend use of createEnrollment() and updateEnrollment()
may need to be updated as well. This has not been done.