removed GetGroupSubmissions(); replaced with GetSubmissions()
This reuses the GetSubmissions() method for getting both group and
regular student submissions; this was already implemented at the
database level. It has also been updated on the frontend. Note that
the frontend must be careful to construct the input query so that
not both UserID and GroupID are set at the same time, as otherwise
there may not be any matches in the database.

We also added hasGroupAccess() to do a full check for the current
user being admin, teacher, student or member of a given group.
(not sure all these checks are necessary...)