major refactor to through out IAssignment and use proto Assignment
This commit removes the unnecessary conversion from Assignment to
IAssignment. It required a function to guard against undefined
deadlines in the Assignment type. Other than this most changes are
trival use of the methods on Assignment. There is one case where
I added a TODO(hein) that must be checked, because it seems to
expect two different ways to produce the timestamp string.
It is also not totally clear if the timestamp string returned by
the getDeadline() function is the desired format for Autograder.
This should be checked.

Note also that I made some adjustments to TempDataProvider to
continue to allow Assignments in the local context. It is not
clear if these will be useful in the future, but maybe still
can be used for testing. At least now they compile.