simpified getGroupUsers by moving some code to access control
This commit moves some of the code previously in getGroupUsers()
to the validation of a group (len > 0) and to access control
using the isInGroup() and isEnrolled() functions.
This avoids having to pass the current user to the CreateGroup()
and UpdateGroup() functions.
This required some updating of the tests since the CreateGroup()
no longer accepts any user (previously an admin user was in the
context); now instead CreateGroup() requires that the current
user in context must also be enrolled in the course and be member
of the group the user is trying to create.

Future commits may move some of the code in getGroupUsers() to
validation.go, but currently this seems a bit much...