added enum field for group repositories
This commit adds a new field for group repo, instead of reusing
the user repository enum field. Also reorganized the values of
the different repo types, placing user and group last, after the
standard repos. Note that this change prevents using from using
the 0 value for a real repo type since it is problematic to
retrieve 0 values from the database. (Not sure about the reason
at this time...) For this reason, I introduced NONE for the 0 value.