removed blank import in webserver/server.go
Importing with a blank identifier as explicit package name
ensures that the package's side-effects (initialization) is
executed, i.e. its init() function. This is already done in
main.go and only needs to (will) be done once. Also a lint rule
prefers to keep such imports in one place (the main.go file).

This explains more
https://stackoverflow.com/questions/21220077/what-does-an-underscore-in-front-of-an-import-statement-mean