added a command line tool to create a secret_ag_test.go
This tool can be used to create the secret and do the init()
so that the user does not need to import the common package
and remember to implement the init() func.
I envision that this would be called for each lab to traverse
each directory to create the relevant secret file.

Usage of gosecret:
  -path string
     directory to traverse for creating secret_ag_test.go (default ".")
  -secret string
     secret string to insert in the secret_ag_test.go