refactored remaining to allow updating assignments and more
This commit (almost) completes the refactoring of the webhooks.
Now updating assignments can be done from both a push event
or from the frontent UI. It required a bit more involved reorg
of the code, adding a new package assignments, but hopefully
this becomes a "little" more clear.

Two small additional changes involve checking that the push
is against the master branch, and only then run assignment update
or run tests. This will allow students to work on a separate branch
without triggering many extra builds until they are ready to
have their code evaluated. We must document this, and encourage
students to work on a branch.

Further improvements can also be made to naming of things in
the new assignments package.