fixed main go.mod file to depend on kit as a module
Now aguis depends on kit as a module instead of as a regular
package. This fixes a problem the prevented aguis from finding
the latest version of the kit package.
This is useful because now external dependencies such as course
tests don't have to download all of aguis and its dependencies
when running tests, but can download only the kit module.