added hack to fix deadlines with wrong format coming from database
Turns out that several courses are using the wrong format, which
has since been stored in the database. This commit runs the
FixDeadline() function also for dates retrived from the database.
In the future, when all courses are parsed and converted at parse
time, and thus stored correctly in the database, this change could
be removed again.