fix bug in processing of assignment strings from github
This commit fixes a bug that was triggered when students push
changes that touch root-level files, that don't belong to any
assignments, e.g. go.mod or README.md and so forth.
The commit also includes a test case for the function that
process the data coming from github.

Note: it is sometimes easy to forget to return or continue when
processing an error. Should be on the lookout for such mistakes.