fixed gogo.proto import and added path in settings.json
This removes the TODO related to gogo.proto import and
adds a general path to the settings.json file; this
settings.json would need to be updated if we update the
gogo package, but the Makefile knows how to find the
most recent gogo package, so this is just a vscode
editor convenience anyway.

Also wrote a bit more extensive doc about the gogo tags.