check if stdout is empty before continuing on error
Some applications (e.g. golangci-lint) send the output to stdout
rather than stderr on an "error". In such cases it's desirable to
compare the error output with the expected output.

This fix uses the contents of stdout to determine if the test failed
in the case stdout has any content.