refactored RunTests to simplify testing with real repos
This is the first of several in a multi-stage commit.
The main refactoring done here is to split RunTests into one more
function runTests() that is suitable for testing without database
access.
The second, less important change in this commit is renaming the
AssignmentInfo.Language field to Script. This change does not
impact the database entires (derived from Assignment.Language,
based on ag.proto), nor does it require changing the yml files
or the assignments package. A future update may change also the
database, yml files, and assignments package.