fixes a performance problem due to always pulling image.
This commit fixes an issue caused by always pulling the image from
docker hub. Assuming that we can leverage the same docker daemon
on the AG server, we can most likely shave off about 30 seconds
from the test execution time. I observed 45 sec test execution
time before this change, and consistently less than 15 seconds
after, when the image is available locally.
This was solved by first trying to create() the container,
and if that fails, then it will try to pull the image, and
retry to create().

The commit also fixes a lint error caused by not handling the
error from ContainerKill().