Return stdout, stderr, error from func CommandLine
Motivation: The output from the CommandLine function is minimal in
some cases. E.g. in our use-case we wanted to detect a data race by
running `go test -race ...` with the CommandLine function and looking
for "DATA RACE" in the output. However we were unable to show the full
output (stack trace etc.) related to the data race when it was found.

With this update the output to stdout and stderr along with any error
from exec.Command are returned as slices, one for each command that
is run. As a result, callers of CommandLine can investigate the
results and use them as they please when some error occurs. Since
there were no return values before, this change should not conflict
with existing code.

I was not sure about introducing new types, but you might also
consider a "CommandLineError" type similar to the following:

```go
type CommandLineError struct {
  // output to stdout
  Stdout string
  // output to stderr
  Stderr string
  // error from executing the command with the exec package
  Err error
}

func (e CommandLineError) Error() string {
  return e.Err.Error()
}
```

Then CommandLine could return `[]CommandLineError` instead.