emit 0 score JSON at start and updated with a non-zero score JSON
This commit fixes #315.

This commit updates the score package in kit module to always
write an initial Score JSON to StdOut. This prevents test score
values from getting lost when students do something wrong which
causes the test execution to panic.

This change implies that Quickfeed must filter the Score JSON
strings found in the results log so that only one Score is
returned and used in the calculation of the Total.
When filtering, we assume only one or two Score strings;
if only one is found, it means that it is the one initialized
by the NewScore* constructor, and it is used, giving only 0 score;
if two is found, only the last one is used, meaning that the
score is the result of running some tests;
if more than two are found, something is amiss and we return
a 0 score object.

This commit also updates the API of NewScore and NewScoreMax.
These functions now take the *testing.T as the first argument,
which is only used to get the test's Name(), which is much more
reliable than our previous approach.

This also removes NewScoreMaxWithTesting() which did the same
and the old testName() helper function, which was unreliable.

Note: This change may not resolve the issue caused by test timeouts,
e.g. due to infinite loops, where there are remaining tests that didn't
get started. Similarly, if a test panics that prevents remaining ones
from even starting, this change will not fix this.