fixed panic handling for subtests and updated docs
This commit adds two new methods:
sc.Print(t): to replace
  sc.WriteString(os.Stdout)
  sc.WriteJSON(os.Stdout)
and
sc.PanicHandler(t) which must be used in subtests to ensure correct
panic handling.

In addition to these new methods, WriteString() and WriteJSON()
have been deprecated and will be removed as soon as the dat320
tests have been updated to use the new API.
@Racin: We must also update the dat520 tests correspondingly.

The commit also updates the documentation for all methods and
the package-level documentation.

Fixes #276