fix for the too many open files problem
This commit fixes #319.
The problem was that we only stopped the container, but didn't
actually remove it, even though we waited for it to finish.
This is mainly an issue resulting from not understanding that
docker containers remains "available" even though the job
running inside the docker container had exited.

We now remove containers that finish without timeout, but only
after having extracted the logs.
Otherwise, we would not be able to get the logs.
If a container times out, we stop the container before removing it.