another attempt at fixing the too many open file descriptors issue
This commit fixes #319 (for real this time).
Now we create only one docker client when we start quickfeed,
instead of creating a separate docker client instance for each
job we want to run. This should hopefully save even more resources
and maybe some execution time for tests.

If this does not work correctly for some reason, we can revert to
the previous version, and instead make sure we close the client
between each job run. But that seems wasteful, and hence I wanted
to reuse the client across jobs. The way I understand the client,
it can still create, stop, get logs from and remove several
containers; that is, we shouldn't need a separate client for doing
these tasks.