refactored group vs user repo selection and test runner
This moves the check for group vs user repo to the main handler
leaving the runAssignmentTests() to do just the actual preparing
and running of tests for a given assignment-repo pair.
This eliminates the need for an addition check in the func, since
we already know that it is either a group or user repo.