remove global secret variable and use environment variable instead
This commit removes the GlobalSecret field, replacing it with a
private variable sessionSecret that then gets initialized in an
init() function via the QUICKFEED_SESSION_SECRET environment
variable. The init() function also unsets the env variable so
that it can't be read by other code after it has been set.

This means that we can eliminate the gosecret command line tool
and thus can update the docker image to remote gosecret.

This is still a draft, and will need more testing when the
lab projects end later this semester.