Started to clean up doc/docker.md
This description needs to be fixed and cleaned up.
However, I stopped since I wonder if this stuff is relevant
with recent Docker packages; seems akward that there wouldn't
be an installer or something that takes care of creating the
docker group, and adding the relevant user (as a default).