Migrated score.proto to score package and added test registry
This is the beginnings of the test registry approach, requiring
the test name, max and weight to be specified before tests
are executed. The objective of this redesign is to prevent that
flawed student code causing test execution to abort early, still
get scored correctly. By registration of all tests before
execution, QuickFeed will already know about all expected tests.