Clarified some docs and made var names camelCase
Mostly nitpicking, but we should be consistent in naming variables
following Go's idiomatic style of using camelCase. When emitting
test error output we should usually follow the style of printing
the func being tested, with which params, what was the result and
wanted/expected result. Sometimes we use 'expected' as well, but
in this file, it looks like we have mainly used 'want' and so
probably good idea to be consistent with that.

Here are two good resources for writing tests:
https://github.com/golang/go/wiki/TestComments
https://golang.org/doc/code.html#Testing