Revised the score API again
Now we use a score registry that doesn't use a singleton map with
score objects. Can now create multiple registries if needed,
although this isn't expected for a single test execution.
Furthermore, the score objects can now also be recorded in the
Scores type, over which we can invoke Sum() and Validate().