Started conversion to new Score type in QuickFeed
This starts the conversion for QuickFeed to use the new Score API.

The Result struct contains Scores and BuildInfo, which are currently
stored in the database and sent to the frontend as JSON strings.
These structs should be translated to protobuf to simplify things.
Make this stuff a separate proto file; maybe move the Result type to
the kit package??
Or maybe not, since BuildInfo is specific for QuickFeed??