Adds GetSubmissionStream
This adds a new RPC func GetSubmissionStream that can be called
to receive one submission at a time in real-time; this could be
called from a client that wishes to update the results as they
arrive to the server. This has only been implemented on the
server-side in ci/run_tests.go:recordResults() and in
web/autograder_service.go:GetSubmissionStream().

This PR has not been tested at all. It would be nice to add a Go
client test function for this that can test the server-side logic,
in addition to support in the web frontend and swift UI frontend.