Change Results message to use slice instead of map
This changes the Results message in score.proto to use a slice
instead of a map to make it easier to work with with respect to
database storage and frontend iteration over the slice.
The scores recorded in the slice is in iteration order of the
respective test execution as before, and so we can avoid using
the previous TestNames slice to define the iteration order.