Change Results message to use slice part for frontend
This updates the relevant frontend code to use slice or list for
the scores, avoiding to iterate over the test names to produce
a list. Now the list of scores come directly from the Results
message type of score.proto.

Further, this commit also reduce some fields of ISubmission to
use directly the BuildInfo message of the Results message.
In addition, a few other updates were made.