Upgraded docker to v20.10.7 and added build image
This upgrades docker, requiring that a few of our uses had to
change accordingly due to API changes in docker's SDK.

This also adds support for building a new image if the requested
image isn't locally available and cannot be pulled from docker.io.

Should also fix pulling images from docker that was previously not
working since we tried to use the PullImage() without specifying
the docker.io/library prefix to the ref string.

Fixes #400 (have only tested locally, not in production).

Fixes #287