Reorganized update.sh script to be more robust
This reorganizes the update.sh script to make backups of the
quickfeed binary, the database file, and the log file before
doing anything else.

Further the script tries to update the code from github and
recompile before checking if server is running; killing a running
server, finally restarting the server and checking whether or not
the restart was successful.