Add Errors slice to score.Results
This commit changes the signature of score.ExtractResults() to
return just a score.Results object that may also contain one or
more errors in the Errors slice. This allows us to log the errors.

This also removes the score.Results message type from the proto file,
instead replacing it with the corresponding score.Results Go type.
That is, the Results type is not stored in the database, and is not
used in the frontend, hence it doesn't need to be in the proto file.