Added dev/package-lock.json and ignore dist/bundle
The package-lock.json is recommended to be added to git.
Updates ui target to use npm ci instead of npm install.
The latter is recommended since it will not install stuff, but
fails with an error if package.json and package-lock.json are
out of sync, allowing dependencies to be updated manually.