Revised Assignment and Task messages in proto file
This commit removes the Issue message and integrates the relevant
fields into the Task message. This also revises the Assignment
message content slightly. I removed the SingleTask field, which is
not needed in my vision of this setup. Also reorganized the ordering
and number of the fields. Hopefully this won't cause problems for
database migration, but all field names are still the same for those
that still exists.