Refactored to use new readTestsRepositoryContent
This commit is a major rewrite of the previous parseAssignments
function. We now first read all the files, then we process all
the assignment files, following by the remaining files. This is
done this way because the "other files" are typically recorded
as part of the pb.Assignment struct, and thus we simply add those
files to its associated Assignment object; for example, the
script template run.sh of an assignment can simply be added to
the assignment.ScriptFile field. Similarly for the criteria.json
which is added to .GradingBenchmarks and task-*.md files are added
to the .Tasks field.