Added pb.Repository.Name() helper method
The Name() method returns the name of a repository, extracted from
the URL path. Note that it is not robust to other URL types than
the ones used by QuickFeed; specifically it will only work with
the last entry of this form: https://github.com/org/reponame,
where Name() will return reponame.