Fixed failing TestGormDBSynchronizeAssignmentTasks
The reason for the test failure was the the inputs to the
SynchronizeAssignmentTasks method come in the form of a map
and map iteration order is non-deterministic in Go. Thus, the
solution is to add the tasks to a slice and sort the slice before
creating the tasks in a deterministic order. Note that the sorting
use the Name field of the task, which should be unique across all
entries in a course.