Refactored TestHandleTasks to TestSynchronizeTasksWithIssues
This is a rewrite of the testing of the task to issue synchronization.
It currently only tests adding new tasks and checking that these
are then reflected as issues on github. The test deletes any remaining
issues on the user repositories before creating new issues. The
created issues are left behind for manual inspection.