Simplified createPullRequest and revised findIssue
Simplified the createPullRequst() method to accept the associated
issue as input; instead moving some of the logic to findIssue()
function that now accepts the list of issues and the pull request
body, returning the issue that matches the pull request body.

The findIssue() also uses regular expression to accept 'fixes',
'closes', and 'resolves' to be compatible with github's notation.
However, the current reg.exp. requires that the "Fixes #30" must
be the last content of the PR body.