Refactored CreateFeedbackComment to MarkdownComment
This moves the CreateFeedbackComment function to the kit/score
package as a method on the Results type. This avoids a parameter
passing, and keeps code for score computation in the score package.

This also refactors several other aspects of the existing code to
remove several exported functions, and adds a test for the main
MarkdownComment method to ensure that the computation remains correct.