Added support for binding directory to /quickfeed in container
This adds support for binding a directory on the host machine
running quickfeed to be accessed via /quickfeed directory.
This will allow us to git clone student code and test code to the
host machine's /tmp directory avoiding to expose the course
creator's access token to the docker image (and by extension
malicious student code). The mounted /tmp directory can be
deleted as soon as the docker container exits, or alternatively
in the future we could maintain a local working copy that can
be synchronized with git pull instead of a full clone...
However, the latter will require more work and does not seem
worth it at this point, given that most repositories are quite
small and quick to clone.