Fixed CreateAssignment and UpdateAssignments for timestamppb
To support recording timestamppb in GORM we cannot use the
Assign(map[string]interface{} ... approach that we previously
used in CreateAssignment. This commit also makes the GetBenchmarks()
method use the struct approach instead of the string-based query
approach; this is strictly not necessary, but hopefully it will
help to detect if a change is needed in the database code, when
refactoring the struct types.