Cleaned up checking of access control methods
This checks that the access control table has an entry for all
service methods. Similarly, it checks that if a service method
is removed without being removed from the access control table.

Running the tests should flag such issues, but if tests are not
run before starting the quickfeed server, it will not run.