Generated grpcweb code with v1.3.1 with javascript disabled
This code was generated without the js_out enabled since the latest
protoc v21.5 release no longer supports javascript:

This line in the Makefile was disabled:
	--js_out=import_style=commonjs:$(1)/proto \

This still compiles since the existing javascript files are still
present in public/proto/qf. However, it should be fixed soon so
that we can again generate using js_out since otherwise we cannot
change the js code, when necessary.