Rearchitected TokenRefresher; use map: method names to func
This commit restructures the TokenRefresher to use a map from
method name to a token updater function. It accomplishes this by
defining two local interfaces that the relevant types must
implement to fit with the general token update function:

func(context.Context, *auth.TokenManager, userIDs) error

The two interfaces are:

type (
	// The userIDs interface must be implemented by request types that may need to update the tokens.
	userIDs interface{ UserIDs() []uint64 }
	// Marker interface to detect the GroupRequest type needed for DeleteGroup.
	isGroup interface{ GetGroupID() uint64 }
)

The userIDs interface must be implemented by all types that are
associated with a method entry in the map; the single user type
simply returns a slice with a single userID.

The isGroup interface is already implemented by the GroupRequest
type used by DeleteGroup; hence, we can use this interface in
a type assertion to ensure that we can call GetGroupID() without
actually using qf.GroupRequest.