Reorg CheckAccessMethods checkAccessControlMethods
This reorganizes the tests used to check that access control is
covered appropriately. Since we no longer want to call the
CheckAccessMethods when starting the server, we only need it to be
unexported and part of a test. That is, we expect a missing
access control entry would cause this test to fail, and thus it
is unnecessary to call it at server startup.