Reorganized MockQuickFeedServer to create QuickFeedService
This renames StartGrpcAuthServer to MockQuickFeedServer and
changes the API slightly to create the QuickFeedService inside
since it represents the server side and isn't supposed to be
called directly. Instead calling the server should happen through
the client interface. Sadly, we still need to pass in the logger
since some of the interceptors use the logger, and thus it is
needed in the tests themselves.