Separate dockerfile build from running tests
This fixes the issue with rebuilding dockerfiles on changes;
we basically remove an existing image and build it again.

That is, we rebuild the image when updating in the repository:
 tests/scripts/Dockerfile

And when RunTests() is called on updates due to push events etc
we do not rebuild the image, but assume that it has already been
built or can be pulled from a remote repository.