Added TestParseOneBadAssignmentAmongCorrectOnes
This test is to show that we get an error if there is one or more
assignment.yml files in the tests repo contains errors.
That is, if readTestsRepositoryContent returns an error, we will
not try to update the assignments in the database.

Possible errors include badly formatted deadline or missing the
assignment's order field. It will still parse without errors and
ignore fields that are not used by quickfeed, such as title and
hours expected that may be used by other tools, such as for
generating the readme.md files for the labs.