refactor(approvelist): major refactor
I got carried away and ended up restructuring much of this tool.
Now there is a struct approveSheet that encapsulates the excel
sheet data and provides helper methods to query the sheet for
students and student IDs etc. More could be done to simplify the
main processing loop be easier to read, but this will do for now.