fix: avoid duplicate autoapprove
For assignments with autoapprove it may happen that students
submit a solution that passes to both their student repo and their
joint group repo, which caused the approval to be counted twice.

This commit ignores submissions from a group repo on a non-group
assignment to be approved, and thus won't be counted. Similarly,
submissions from a student repo to a group assignment will also
be ignored.

This is mainly an issue for autoapprove assignments.