refactor(scm): create issues instead of populating mock SCM
This removes the initialization of the s.issues map; instead we
rely on creating issues prior to getting and updating those issues.
This should make the usage of the mock http handler more flexible.
Note that we still initialize organizations and repositories since
issues are associated with a repo. Issue comments must be assoicated
with an issue, and thus also requires creating issues before
creating and updating an issue comment.

This also adds the hasOrgRepo helper function to check if a given
organization and repository exists in the orgs and repos slices.