test(scm): added TestMockUpdateEnrollment and mock http handler
This adds a new version of TestMockUpdateEnrollment that will
replace the old version. This should cover more cases and actually
tests the GithubSCM implementation.

Note this adds the API routes for deleteOrgs and deleteRepos, but
we still need to add the actual handlers for those in a follow up
commit.