test(scm): added TestMockRejectEnrollment and mock http handlers
This should replace the previous TestMockRejectEnrollment impl that
rely on MockSCM.

Note that RejectEnrollment may delete a user from an org and then
fail to delete the repo; thus, RejectEnrollment is not idempotent.
We may consider to log the error if the user has already been
deleted and try to delete the repo anyway. But perhaps it makes
sense to leave the repo around, and we could do manual clean up
in this case. Not sure which is best.